Reorganize YaspGrid constructors
Until now, there were 6 constructors for a YaspGrid. Using the
fact, that CollectiveCommunication is (or will be) default-constructible,
we can boil this down to 2 constructors - one for equidistant, one
for tensorproduct grids. 4 of the old 6 constructors are backwards compatible
and are immediately deleted. the other 2 are deprecated. The methods
<Equidistant,TensorProduct>Setup are moved into the respective constructors.
For the time of the deprecation, this code is duplicated (alas the line
count of this commit).