[Ranges] Make ranges only available if range-based loops are supported
Originally, this patch set tried to make the range generating functions
available to all supported compilers, as they might be useful even
without range-based for loops. Unfortunately, GCC 4.4 has a serious
compiler bug that makes it impossible to backport the
functionality. Instead of introducing an additional check that would
make the functions available on GCC 4.5, this patch simply hardcodes
range-based for loops as a requirement.

The docs have been adjusted accordingly and now state that GCC 4.6 is
required for the range generating functions.