[Ranges][Tests] Clean up test for iteration ranges and remove fallback support
This patch separates the actual tests from the grid creation process,
making it easier to perhaps extract those tests later on if people want
to run them on their own grids.

It also removes the fallback code for compilers without range-based for
loops as well as the backwards compatibility binary that was built to
test the fallback code on newer compilers.