Deprecate the values Geometry::dimension and Geometry::dimensionworld
As decided at the 2014 dev meeting in Berlin.  Both values convey information
about the grid that a geometry may be part of, but not about the geometry
itself.  Therefore they should not be part of the Geometry interface.

Unfortunately, the deprecation here is only a deprecation in the doxygen
documentation.  No other form of enum deprecation appears to work.