Rename the template parameter of YaspGrid and some implementation classes
Originally the coordinate type parameter was called CoordCont or CC.
As on the developer meeting in Berlin, it was decided to drop the "Container"
from the naming scheme, this should now be "Coordinates" instead of
"CoordCont". The choice of the acronym CC has always been unlucky: It was
used for CoordinateContainer and CollectiveCommunication alike. All
template parameters CC have been renamend to Coordinates too.