[bugfix] Correctly restore grids with varying refinement option
The option, whether the overlap should be kept in physical size
or in number of cells can be set individually on each refinement
of the grid. The facility assumed the parameter to be constant
over all refinement steps. This patch fixes this issue, by storing
the refinement option of each grid level.