Implement a GlobalIndexSet class
The class implements an index set over all entities in a distributed grid.
This patch introduces the code as presented in FS 1489, with minor modifications.
It is fully functional (as far as I can tell).

Future patches will improve the interface.

Thanks to Benedikt Oswald and Patrick Leidenberger for donating the original
version of this code.