[YaspGrid] Make YaspGrid entities copyable and movable
With this patch, it becomes possible to copy, move and store YaspGrid
entities.

The old entity implementation relies on an EntityPointer to provide
storage for the data required by the entity; the entity then simply
contains pointers to this data. In particular, an entity object contains
a pointer to the containing grid as well as pointers to two objects
currently stored in the EntityPointer:

- an iterator that provides access to the grid level
- the actual YGrid iterator that is used to iterate over the entities

In order to make the entity independent of its underlying
EntityPointer, this patch moves those two objects into the entity itself
(with the nice side effect of improving readability of the code, as
those pointers don't have to be derefenced all the time anymore).

The EntityPointer now only contains an entity object. Note that in
contrast to the old implementation, it actually contains the full facade
object instead of the implementation object. That way, we don't have to
construct a temporary every time the iterator is dereferenced, which
also used to force a copy of the implementation object. That copy has
just gotten more expensive (all the data now lives inside the entity),
so it's a good idea to avoid the copy.

We can also get rid of all the custom copy code for the EntityPointer,
as the autogenerated ones will now do the right thing for the shiny new
entity stored inside the EntityPointer.

Finally, we have to modify both the LevelIterator and the
HierarchicIterator, which used to operate on the data members of the
EntityPointer base class. As those members are now stored directly
inside the entity, the iterators need access to the entity
implementation, so they had to become friends of both the grid (so they
can use Grid::getRealImplementation()) and the entity.