[Interface] Deprecate EntityPointer
EntityPointer will be removed after the release of dune-grid-2.4 and
will be replaced by the ability to copy and store entities directly.

This patch deprecates the EntityPointer facade and most of its interface
methods. It also adds explicit overrides for the dereferencing operators
to the EntityIterator facade to hide the deprecated versions in the
EntityPointer base class.