[Interface] Make Entity dereferenceable for backwards compatibility
When changing interface methods that currently return an EntityPointer
to instead return an Entity, user code shouldn't directly break.

This patch works around the problem by adding operators to the Entity
facades that return references to itself. To make the user aware of the
problem, those new operators are deprecated.