[Interface] Add deprecation warning infrastructure for EntityPointer usage
This patch adds a utility method called warnOnDeprecatedEntityPointer()
to the codim 0 entity facade. This method takes a single template
parameter which has to be a facade class of either an Entity or an
EntityPointer for the current grid.

If the template parameter is an EntityPointer, a deprecation warning
will be generated.

The two versions of the functions are realised with SFINAE because it is
not possible to specialize member functions of a template class. In
order to avoid a warning, the implementation *MUST* return the entity
after wrapping it in the facade!