[Interface] Replace EntityPointer with Entity if supported by implementation
This patch changes all methods in the grid interface that currently
return an EntityPointer to return an Entity instead:

- Entity<0,...>::father()
- Entity<0,...>::subEntity()
- Intersection::inside()
- Intersection::outside()

Performing this change unconditionally would break all grids that have
not been converted to return an Entity from the corresponding
implementation methods. To avoid extensive breakage,the facades
automatically fall back to returning an EntityPointer if the
implementation has not been converted yet.

In order to simplify access to the types of subentities in the codim 0
Entity, the patch also adds an Entity typedef to the nested Codim struc,
alongside the existing EntityPointer typedef.

WARNING TO GRIDIMPLEMENTORS:
This fallback test relies on the implementation returning a facade
object. If your grid has been returning an implementation object and
relying on the interface to wrap it correctly, this patch *will* break
your grid!