[YaspGrid] Return Entity instead of EntityPointer from interface methods
This change is pretty straightforward: In the Entity methods we just
construct an Entity instead of an EntityPointer. In the intersection
life is even easier because we already switched to storing entity
implementations.

The only thing to be aware of is the additional call to the explicit
constructor in the facade.