[Interface] Make Entity and EntityPointer comparable
In order to improve interoperability between existing code and grids
that return entities instead of EntityPointers, this patch makes it
possible to directly compare those two kinds of objects.

The comparison operators are deprecated to let people know that those
comparisons are evil. But with those changes, the gridcheck for YaspGrid
compiles again (albeit with insane amounts of warnings...).