[Yaspgrid] Introduce a coordinate object for equidistant grids with non-0 origin
As discussed on the mailing lists it would be nice to have equidistant grids
with a user-defined lower-left corner without resorting to a tensorproduct
grid. This commit introduces a coordinate object called
EquidistantOffsetCoordinates<ct,dim> which implements the desired behaviour.
To use it, an additional constructor is introduced, which takes the
lower-left and upper-right corner of the cube as first and second parameter.

Performance measurements proved that such new coordinate object is necessary,
because a slowdown of roughly 10% is observed for normal equidistant grids
having a 0 origin.