Rename method 'nGlobalEntity' to 'size'
'size' is the name the method has in the dune-grid IndexSet class.
The IndexSet class method also has a parameter 'codim', and returns
only the number of entities of that codim.  The GlobalIndexSet only
stores indices for one codimension anyway, so the 'codim' parameter
would be pointless here.  We add it nevertheless, for increased
consistency with the dune-grid IndexSet class.