Move the specialization of StructuredGridFactory for SGrid from structuredgridfactory.hh to sgrid.hh
Specializations for specific grids should not be in the file structuredgridfactory.hh,
because then the complete headers for those grids are included everytime the header
structuredgridfactory.hh is included.  For example when the StructuredGridFactory class
is used with a UGGrid, the complete implementation for SGrid and YaspGrid is included.
This will not help your compile times.

Additionally, since sgrid.hh is deprecated now, users get the deprecation warning
every time they include structuredgridfactory.hh.  In particular even when they
are not using SGrid at all.  This patch solves the issue by moving the specialization
StructuredGridFactory<SGrid> from the structuredgridfactory.hh header to the sgrid.hh
header.

This change is completely backward-compatible, because people that do use SGrid need
to have sgrid.hh included anyway.