[Interface] Deprecate comparisons between different iterator types
As discussed in FS#1513, comparing different types of iterators is at
best an exercise in ambiguity, and its main use case disappears with the
removal of EntityPointer.

So this patch deprecates those kinds of comparisons. Comparing two
iterators of the same type is now handled by two new non-template
comparison operators. Moreover, there are now also new comparison
operators for EntityPointer because the new templated operators shadow
the versions in EntityPointer itself.