[cleanup] Make UGGridGeometry return matrices by value instead of by const-reference
This reflects a corresponding recent switch in the Geometry interface class.

Returning matrices by value allows to get rid of the internal copies of the matrix.
This in turn forces us to give up on matrix caching for simplex elements (or make
the caching more complicated).  In this patch, caching is removed.  It makes the
code simpler; I didn't test potential impact on run times.