[Gridcheck] Adapt to copyable entities and intersections
This patch updates the grid check to use copyable entities and
intersections and to support grids that return entities instead of
EntityPointers from interface methods like father().

As the interface change is fairly intrusive (we need to change every
location where subEntity(), father(), inside(), outside() or
Grid::entityPointer() are called), it is not possible to "outsource"
this change into a separate check that grid implementors can deactivate
if their grid does not yet support the new interface. At the same time,
we want backwards compatibility in 2.4, so people should be able to use
the return values from the functions listed above like an EntityPointer.
As a consequence, we have unfortunately have to introduce dozens of #ifdefs
in all kinds of places.

The patch introduces a new preprocessor define called
DUNE_GRID_CHECK_USE_DEPRECATED_ENTITY_AND_INTERSECTION_INTERFACE. This
define must be set when running the grid check on a grid that has not
been ported yet. It will generate a large amount of deprecation
warnings, irrespective of whether DISABLE_DEPRECATED_METHOD_CHECK has
beet set or not because it forces the check to use EntityPoitners, and
those are deprecated.

The general idea behind most of the fixed locations is this: If the grid
is either not ported yet or if deprecated methods are checked, the
function return value is captured in an EntityPointer which is
afterwards dereferenced to avoid unused variable warnings. For unported
grids, we then create a variable that holds a const ref to the entity
stored in the EntityPointer. For grids that have been ported, we instead
call the interface method again and store the result as a copy in a new
variable. After all this setup, the following code is changed to use the
new variable instead of always dereferencing the EntityPointer.

In some places, the patch duplicates longer sections of code because it
would otherwise have become completely unreadable (e.g. in the
geometryInFather check).

Finally, comparisons between different types of iterators and / or
EntityPointers are only checked when deprecated method checks are
enabled anymore.

I also fixed a small bug in assertNeighbor() where a check on
LevelIntersectionIterators was indadvertently disabled when deprecated
methods checking was turned off, as someone just switched off a call to
Entity::ilevelbegin() instead of porting it to GridView::ibegin().