[GeometryGrid] Add equals() methods to entities and intersections
This patch adds equals() methods to entities and intersections to allow
comparisons via the facade classes.

They are mostly just ported over from EntityPointer and
IntersectionIterator, with an additional check for pointer validity.

The comparison semantics in the case of synthesized entities aren't
entirely clear to me; I have opted for first checking the subEntity
member (as done by the EntityPointer) and then checking whether both
pointers to the host entities are valid. If they are not, two entities
compare equal iff both pointers are invalid. Otherwise, we can proceed
with the comparison as copied over from the EntityPointer.