[UG] Add new constructor to create entity from target and grid
This patch adds a new constructor to directly initialize an entity with
a target pointer and the grid to be able to directly return an
initialized entity from methods like father() and subEntities when
switching those over from EntityPointer to Entity.

This functionality is already present on UGMakeableEntity, but as
UGMakeableEntity is a subclass of the actual entity returned to the
user, its not a good idea to use it in this context, as the required
slicing to the base class will always force an extra copy and possibly
inhibit return value optimization.