[UG] Return entities instead of EntityPointers
This patch changes all interface methods that used to return an
EntityPointer to instead return an Entity.

It also adds a new method entity(const EntitySeed&) to the grid which
returns the entity for a given EntitySeed and deprecates the older
method entityPointer(const EntitySeed&).

Finally, it switches the gridcheck over to the new interface.