[Bugfix][YaspGrid] Fix compilation problem with GCC 4.4
GCC 4.4 has a number of nasty bugs in its rvalue reference support. As
reported by Markus, it fails to disambiguate between two
YaspGridEntityPointers in this case, so this patch disables the
optimized (rvalue ref) version.

As this behavior is hard to trigger in small configuration tests and
because we'll probably drop GCC 4.4 after the next release anyway, I
hacked in this hardcoded check for the compiler version.

Fixes FS#1534.