Introduce methods 'index' and 'subIndex' to eventually replace 'map' and 'map'
At the 2014 dev meeting in Berlin it was decided that the method 'map' of
the Mapper classes should be renamed to 'index' and 'subIndex', respectively,
to be more consistent with the IndexSet interface.  This patch does the
renaming.  The old methods are still there, but they are now marked as
deprecated.