Introducing ParMetisGridPartitioner, a helper class to partition Dune grids with ParMetis
This patch adds a new file with one class ParMetisGridPartitioner.  This
class provides a convenient way to obtain grid partitionings using ParMetis
for arbitrary Dune grids.

Using only methods of the grid interface, it assembles the adjacency information
required by ParMetis, and then calls ParMetis to obtain an initial partitioning
or a repartitioning.  The output is given in a std::vector<unsigned int>, which
for each interior element contains the number of the process where this element
should be sent to.

Currently, there is no standardized why to feed repartition information into
a Dune grid.  Using std::vectors here is a proposal, and I will not object to
changes if somebody comes up with a smarter idea.  Currently, the output of
the ParMetisGridPartitioner class can be fed directly into the loadBalance
method of UGGrid, so now you can repartition UGGrid objects with ParMetis.