[YaspGrid] Drop YaspGrid pointer from entities
The YaspGrid entities store a pointer to their YaspGrid, but that
pointer is only ever used in the codim 0 entity to access the maxLevel()
of the grid in isLeaf() and isNew(), so carrying those extra 8 byte
around all the time seemed a tad excessive.

So I've put the little beasties on a diet and removed the pointer. The
YaspGrid is still accessible through the YGridLevelIterator, and while
that takes an extra indirection, it's so rarely used that it really
doesn't matter.