[fix future bugs] Add an abstract base class for YaspGrid load balancing objects
Problem: The class YLoadBalance defined the interface for Yasp load balancing
objects AND provided the default implementation. Since the last release, the
signature of its loadbalance method has changed due to the FieldVector<int> -> array<int>
change. User code that implemented its own YLoadBalance object by deriving from
it is now ignored, if the signature not adjusted. This is a severe bug that the
compiler is not able to warn about.

Solution: Make YloadBalance an abstract base class. The default implementation
is moved to a new class YLoadBalanceDefault. This way, user provided load balancing
objects are forced to implement the correct signature.