[YaspGrid] Put all the partitioning related stuff into a separate header
Users that want to implement their own partitioning strategy
need a place to look for how it is done and what strategies are
already available. A separate header is the right place for that.
The partitioner from backuprestore.hh, which takes an array with
a number of processors per direction, is of general interest (in
fact many users have implemented the same thing). It is also moved
there and renamed to YaspFixedSizePartitioner. The naming on this
one differs greatly from the others, but that is because I aim
for renaming the other ones in 3.0 (its partitioning and not
loadbalancing what we are doing, after all).