[GeometryGrid] allow access to coordinate function
This patch allows the user to access the coordinate function of a
GeometryGrid after its creation. This is a pure convenience method,
as GeometryGrid only stores a reference to the actual function.

Notice that the type of CoordFunction has always been exported publicly.
This is hidden in the class GeoGrid::ExportParams.