[CMake][Bugfix] Fix FindALUGrid.cmake in the parallel case
The test fails if the METIS/ParMETIS dependency is
not correctly handled. The test ignored this up to now.
It now calls find_package([Par]METIS) when ALUGrid with
parallel headers was found. The METIS libs are added to the
alugrid libs. Similarly, the add_dune_alugrid_flags macro
now also calls the corresponding metis macros.