[IdentityGrid][1/11] Port to copyable entities and intersections
This patch prepares the EntityPointer implementation for the copyable
entities. Instead of storing a mutable entity, it now contains a host
EntityPointer and returns a temporary entity when
dereferenced. Moreover, taking inspiration from GeometryGrid we now
directly store the most-descended host iterator type instead of a plain
host EntityPointer. The type of the host iterator is passed as an
additional template parameter. This tricks avoids storing two separate
host EntityPointers for iterators.