[IdentityGrid][2/11] Port to copyable entities and intersections
This patch ports the iterators to the new EntityPointer
implementation. We now

- store the host iterator in the EntityPointer base class
- directly advance that host iterator in increment()
- don't have to care about updating a stored entity anymore!

Taken together, these chagnes remove all data members from the
iterators, which was also a chance to drop weird stuff like the fact
that the leaf and the hierarchic iterator (but not the level iterator)
always stored an additional end iterator that was never used and some
useless additional pointers to the IdentityGrid.

Finally, all occurences of deprecated lbegin() / lend() / leafbegin() /
leafend() calls have been replaced by calls on the respective
GridViews.