[IdentityGrid][3/11] Port to copyable entities and intersections
For our EntityPointer optimization trick of directly storing the
most-derived host iterator, we had to change the type signature of the
implementation and thus have to somehow integrate that change into the
GridTraits.

This patch takes care of that problem by letting the actual Traits
inherit from the default implementation and fixing the EntityPointer
typedef.