[IdentityGrid][8/11] Port to copyable entities and intersections
This patch cleans up the intersection iterators. Instead of the
previous, backward setup where the actual host iterator was stored
inside a mutable IdentityGridIntersection, the iteratos now simply store
a host iterator and return temporary intersections when dereferenced.

This patch adds an IdentityGrid pointer to the iterators because that
pointer has to be passed to the newly constructed intersection objects.