[Interface][Bugfix] Make Entity::father() work for GCC < 4.8
Older versions of GCC are a little over-eager when trying to evaluate
the return type of the non-template member function father(). In
particular, they try to resolve the decltype() type expression in the
return type, which must only happen at template instantiation time.

This patch avoid the problem by turning father into a template with a
default template argument, which (together a little SFINAE) reliably
delays return type determination on GCC 4.4+ to function instantiation
time.

In order to avoid confusing average users, the modified signature is
hidden from Doxygen.