[Interface][Bugfix] Make new Entity / EntityPointer switch work with GCC 4.4
The new switch in the interface that provides automatic backwards
compatibility with grids that still return EntityPointers from methods
like father() etc. has a problem with GCC 4.4 and the subEntity() method
of codim 0 entities because that compiler doesn't correctly apply
visibility rules during return type deduction (it uses the scope of the
caller instead of the callee when applying visibility
constraints). Moreover, that compiler also has an incomplete decltype()
implementation that sometimes causes an internal compiler error when
decltype() is used during return type deduction.

This patch works around the problem by moving all of the type deduction
into a separate, dedicated helper struct, which avoids triggering the
decltype() problem. Moreover, instead of simply accessing the realEntity
member of the Entity, we now construct a bogus pointer of the correct
type in the decltype expression. To top it all of, GCC 4.4 doesn't even
have std::declval(), so we have to do even that manually...

Sigh, good thing that that compiler will be gone after DUNE 2.4.