[YaspGrid][Bugfix] Fix compilation problem reported in FS#1534 for entities.
GCC 4.4 sometimes fails to disambiguate correctly between overloads for
const lvalue references and rvalue references. This has already created
problems for the EntityPointers (see FS#1534, fixed in 2b24a516), but a
similar problem now crops up for the copyable entities, so this patch
applies the same workaround to the entity implementation.