[GeometryGrid][4/12] Implement copyable entities / intersections
This patch makes GeometryGrid entities copyable, which involves two
separate concerns:

1) Move all required data into the entity class, so that an entity is
   self-sufficient and does not require an accompanying EntityPointer.
2) Add new copy and move constructors. In addition, this patch also
   adds a default constructor, making it possible to default-construct
   invalid, but consistent entities that can be assigned a valid entity
   later on.

Instead of storing a pointer to a host entity (which is supposed to be
contained in the host iterator of the accompanying EntityPointer), the
entities now directly store a host entity. In addition, they also store
a pointer to the GeometryGrid, as they need to be able to provide that
pointer to the geometry implementation once that gets
constructed. Accordingly, the grid() method no longer forwards to the
geometry, but returns the internal pointer. I also removed a number of
unused constructors and the initialize() methods, which are not needed
any longer.

WARNING! As I don't have a grid with missing entities for some
codimensions around, the support for emulating those missing entities
may be broken!