[GeometryGrid][5/12] Implement copyable entities / intersections
This patch adapts the EntityPointer to the new, self-contained
entities by removing the contained mutable entity copy.

GeometryGrid now returns temporary entities from its
EntityPointers/iterators and which contain a copy of the host entity.

As the GeometryGrid entities require a pointer to the grid, the
EntityPointer now also stores such a pointer.

I tried my best to port the version of the EntityPointer that uses the
emulated entities, but as for the entities itself, without a host grid
to test the feature, I don't know whether it works. YMMV.