[GeometryGrid][8/12] Implement copyable entities / intersections
This patch makes GeometryGrid intersections copyable. The overall idea
behind the patch is similar to patch 4 in this series, which does the
same for entities: Instead of just storing a pointer to the host
intersection stored inside the host IntersectionIterator of an
associated GeometryGrid IntersectionIterator, we now simply store a host
Intersection. This change also gets rid of the initialize() /
invalidate() logic and allows us to rely on the default copy
constructors and assignment operators as well as their move equivalents.

The patch also allows intersections to be default-constructed into an
uninitialized, but defined state that can be made valid by copying
another intersection into it and simplifies the comparison logic in
equals().