[GeometryGrid][9/12] Implement copyable entities / intersections
Switch intersection iterator to use new self-sufficient
intersections. The iterator now returns temporary copies of
intersections, allowing us to get rid of the internal intersection
variable and the associated initialize() / invalide() logic.

On the other hand, we now need to store the geometry of the inside
entity to pass it to the newly created intersections.