[Gridcheck] Add checks for range-based iteration and entity / intersection lifetime
This patch extends the gridcheck with two new methods that try to
iterate over entities / intersections using the new range-based for
constructs. Moreover, it stores copies of a number of entities and
intersections and then makes sure that those entities and intersections
remain valid as long as their stated lifetime requires.

If the grid does not support copyable entities / intersections, the test
is skipped. Moreover, the test only runs the LevelIntersection tests if
the grid indicates support via EnableLevelIntersectionIteratorCheck.

Finally, as the tests rely on range-based for loops, they are skipped if
the compiler does not support those (currently, this mainly affects GCC 4.4).