[AlbertaGrid][1/12] Implement copyable entities / intersections
This patch implements correct move semantics for ElementInfo and its
reference-counted data pointer.

The idea behind this patch is to add a special case, where instance_
points at nullptr. normally it never does, because that pointer is
dereferenced in a number of places without checking its validity. For an
rvalue that has become the argument of a move operation, this is safe,
however, because the compiler guarantees that the object will be
destroyed directly after leaving the move constructor / assignment
operator. This scheme avoids unnecessary fiddling with the reference
count of the null element in the instance stack.