[VTKWriter] Use new abstraction for VTK data sets based on LocalFunction concept
This patch introduces a new type VTKLocalFunction that provides a
value-based interface to VTK data sets, which are stored using type
erasure. There are currently two implementations of the type erasure,
one for legacy VTKFunctions and one for the LocalFunction concept of
dune-functions.

Internally, an important change is the switch from evaluating
vector-valued data sets once for each component to a single,
vector-valued evaluation, which should help performance for complex
discretizations.

Apart from that, the patch also deprecates the add*Data() functions that
take a naked VTKFunction pointer and cleans up quite a lot of the
internal implementation.