[vtk,bugfix] create a const ref befor assigning to the entity
Depending on the grid implementation, the entity is now returned as a copy.
This implies that we can not call bind(*it), as we bind to a temporary.
We now create a const ref which is valid until the end of the code block
and bind to the const ref.