[Release][GridCheck] Make it possible to skip the BoundarySegmentIndexCheck
Some grids (e.g. the subdomain grids of MultiDomainGrid) are not
expected to contain valid BoundarySegmentIndices, which currently makes
it impossible to run the gridcheck on them without errors.

This patch makes that part of the gridcheck optional by adding a new
feature flag Dune::EnableBoundarySegmentIndexCheck. It is modeled on the
existing EnableLevelIntersectionIteratorCheck, but used the interface of
std::integral_constant (::value instead of ::v) and is placed inside the
Dune namespace.