[Release][Bugfix] Avoid invalid iterator access in YaspGrid
The YGridList, used by YaspGrid to store a nested list of grid
intersectionswith neigboring processes, calculates an offset for the
YGridComponents associated with each element of the outer list. When
updating this offset in the finalize() method, the  implementation does
not check whether the current outer list element is empty. If it is
empty, the code derefences an invalid iterator, causing a segfault when
compiled with Clang 3.6.

This patch fixes the problem by only calculating an offset if the
current list element is non-empty.

(cherry picked from commit f693f563331dcf920dcc803f0cd850eb1c74dcda)
Signed-off-by: Steffen Müthing <muething@dune-project.org>