[release] Stop cmake from looking for UG by consulting the PATH environment variable
The following construction is part of the standard cmake search locations
for find_package: [quote from http://www.cmake.org/cmake/help/v3.0/command/find_package.html ]

"Search the standard system environment variables. [...] Path entries ending in /bin or /sbin
are automatically converted to their parent directories."

This leads to subtle problems building with UG when all of the following holds:

- you want a global UG installation
- you have a local UG source tree in your home directory
  (which contains a ug-config.cmake)
- your PATH contains $HOME/bin

which incidentally happens to be the case on my machine.  cmake then looks for
the package in subfolders of the home directory, and 'finds' it in the source
tree.  Since that is not a real installatioa,n linking fails eventually because
the UG libraries are not passed to the linker.

Setting NO_SYSTEM_ENVIRONMENT_PATH disables this particular cmake behavior.

Thanks to Ansgar for helping me debug this.