Move specialization of StructuredGridFactory for YaspGrid out of the structuredgridfactory.hh header
Previously, the header file structuredgridfactory.hh contained a specialization
of the class StructuredGridFactory for YaspGrid.  For this to work, structuredgridfactory.hh
had to include yaspgrid.hh.  That means that every user of the StructuredGridFactory
had to parse the complete YaspGrid implementation, even if the YaspGrid specialization
of StructuredGridFactory was not used.

This patch moves the specialization to a special header in the yaspgrid subdirectory.
Now structuredgridfactory.hh does not include yaspgrid.hh anymore, but in turn
yaspgrid.hh (implicitly) includes structuredgridfactory.hh.  This is a much
cheaper price to pay.