[cleanup] Simplify reading of element tags
I happened to see that the element reading code in the GmshReader
contain a few lines of dead code.  Each element has a variable-length
list of integer 'tags'.  We read the first of these tags as the
'physical entity'.  The code however, read and stored the remaining
tags too, using some not-quite-evident logic.  Then, before reading
the next element, the stored tags were ignored again.

This patch ignore all but the first task right away, which makes the
code 10 strange lines shorter.