[bugfix][release] fix configuration when MPI is not found
The test target mpivtktest is only added when MPI_FOUND is set. Thus
cmake aborts with an error when setting testproperties for mpivtktest
in case that MPI is not found.
Thus we have to set the testproperties for mpivtktest only when
MPI_FOUND is true.

(cherry picked from commit b117e33ec1671171703c1fc8be49ad0a6de550c0)
Signed-off-by: Carsten Gräser <graeser@mi.fu-berlin.de>