[Bugfix][Release][YaspGrid] Fix index calculation in DataHandle for intermediate codims
YaspGrid stores lists of intersections for each of the possible
communication patterns specified by the grid interface. For
intermedidate codims (0 < codim < dim), there are several of those
lists, one for entity orientation (or shift in the YaspGrid
codebase).

There is a bug in the setup code for those intersections that causes
YaspGrid to incorrectly calculate index offsets for some of those
directions if there are no intersections at all for another direction.

This code fixes the problem, for a more detailed explanation see the
comment at the begin of YGridList::finalize().