[CMake][Bugfix][Release] Don't link libdunegrid to itself
The diverse add_*grid() macros link a target with all of the libraries
required to work with the grid. This naturally includes libdunegrid
itself. Normally, this is exactly what we want, but those macros are
also called to add the external libraries of the grid managers to the
dunegrid library target itself, which then causes libdunegrid to be added
as a link dependency to libdunegrid. This is not allowed anymore
starting with CMake 3.0 (see policy CMP0022).

This patch works around the problem by adding a new flag to the
add_*grid() macros called NO_LINK_DUNEGRID, which causes them to not
link the target to libdunegrid. This special flag is then used when
calling the macros for the libdunegrid target.

(cherry picked from commit d92181a28cc3032b12890df2a4351d71cf443d8f)
Signed-off-by: Steffen Müthing <muething@dune-project.org>