Simplify default implementation of method hasBoundaryIntersections
The previous implementation looped over all level intersections, and over
all leaf intersections, too.  However, if an intersection is on the boundary,
it is so both for the level and the leaf view.  Hence, the second loop
is not necessary.

(cherry picked from commit 536facdd79a9ddb6d188497c5736bb0a3b35c51e)
Signed-off-by: Steffen Müthing <muething@dune-project.org>