add physical entities to gmsh output
The user can now provide a vector of physical entities as a parameter
to the write method. They are stored as the first tag of the taglist of
each grid element. The writer uses the ordering of the index set of the
associated grid view via the MultipleCodimMultipleGeomTypeMapper using
the MCMGElementLayout.