add physical entities to gmsh output
The user can now provide a vector of physical entities as a parameter
to the write method. They are stored as the first tag of the taglist of
each grid element. The writer uses the ordering of the index set of the
associated grid view via the MultipleCodimMultipleGeomTypeMapper using
the MCMGElementLayout.

(cherry picked from commit a7885ca2e1a9ac3e4617279fdc2edc056b98b75e)
Signed-off-by: Steffen Müthing <muething@dune-project.org>