[bugfix] Throw an exception on all processes if the grid is too small
If overlapping computation on a too small grid is requested, an exception
is thrown by the YaspGrid constructors. However, the condition that is
checked for that does not trigger on all processes. The processes where no
exception was thrown did not terminate. This commit solves the problem
by communicating the result of the check to all processes.

(cherry picked from commit a460b2d8576c16e7085c1ddc8e7ee27a83826456)
Signed-off-by: Steffen Müthing <muething@dune-project.org>